Code review changes - Added some tests
