+++ api/rest/docs/swagger.json
@@ -153,7 +153,7 @@
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/core.BidCaching"
+                            "$ref": "#/definitions/dto.BidCaching"
                         }
                     }
                 }
@@ -1862,7 +1862,7 @@
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/core.RefreshCache"
+                            "$ref": "#/definitions/dto.RefreshCache"
                         }
                     }
                 }
@@ -2506,41 +2506,6 @@
                 },
                 "floor": {
                     "type": "number"
-                },
-                "os": {
-                    "type": "string"
-                },
-                "placement_type": {
-                    "type": "string"
-                },
-                "publisher": {
-                    "type": "string"
-                },
-                "rule_id": {
-                    "type": "string"
-                }
-            }
-        },
-        "core.BidCaching": {
-            "type": "object",
-            "properties": {
-                "actvie": {
-                    "type": "string"
-                },
-                "bid_caching": {
-                    "type": "integer"
-                },
-                "browser": {
-                    "type": "string"
-                },
-                "country": {
-                    "type": "string"
-                },
-                "device": {
-                    "type": "string"
-                },
-                "domain": {
-                    "type": "string"
                 },
                 "os": {
                     "type": "string"
@@ -3903,38 +3868,6 @@
                 }
             }
         },
-        "core.RefreshCache": {
-            "type": "object",
-            "properties": {
-                "browser": {
-                    "type": "string"
-                },
-                "country": {
-                    "type": "string"
-                },
-                "device": {
-                    "type": "string"
-                },
-                "domain": {
-                    "type": "string"
-                },
-                "os": {
-                    "type": "string"
-                },
-                "placement_type": {
-                    "type": "string"
-                },
-                "publisher": {
-                    "type": "string"
-                },
-                "refresh_cache": {
-                    "type": "integer"
-                },
-                "rule_id": {
-                    "type": "string"
-                }
-            }
-        },
         "core.RefreshCacheFilter": {
             "type": "object",
             "properties": {
@@ -4181,6 +4114,41 @@
                 }
             }
         },
+        "dto.BidCaching": {
+            "type": "object",
+            "properties": {
+                "actvie": {
+                    "type": "string"
+                },
+                "bid_caching": {
+                    "type": "integer"
+                },
+                "browser": {
+                    "type": "string"
+                },
+                "country": {
+                    "type": "string"
+                },
+                "device": {
+                    "type": "string"
+                },
+                "domain": {
+                    "type": "string"
+                },
+                "os": {
+                    "type": "string"
+                },
+                "placement_type": {
+                    "type": "string"
+                },
+                "publisher": {
+                    "type": "string"
+                },
+                "rule_id": {
+                    "type": "string"
+                }
+            }
+        },
         "dto.BidCachingUpdRequest": {
             "type": "object",
             "properties": {
@@ -4308,6 +4276,38 @@
                     "type": "string"
                 },
                 "user_full_name": {
+                    "type": "string"
+                }
+            }
+        },
+        "dto.RefreshCache": {
+            "type": "object",
+            "properties": {
+                "browser": {
+                    "type": "string"
+                },
+                "country": {
+                    "type": "string"
+                },
+                "device": {
+                    "type": "string"
+                },
+                "domain": {
+                    "type": "string"
+                },
+                "os": {
+                    "type": "string"
+                },
+                "placement_type": {
+                    "type": "string"
+                },
+                "publisher": {
+                    "type": "string"
+                },
+                "refresh_cache": {
+                    "type": "integer"
+                },
+                "rule_id": {
                     "type": "string"
                 }
             }
