+++ api/rest/refresh_cache_test.go
@@ -1,11 +1,9 @@
 package rest
 
 import (
-	"encoding/json"
-	"github.com/m6yf/bcwork/core"
+	"github.com/m6yf/bcwork/dto"
 	"github.com/m6yf/bcwork/models"
 	"github.com/stretchr/testify/assert"
-	"github.com/volatiletech/null/v8"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -14,7 +12,59 @@ import (
 )
 
 func TestValidateRefreshCache(t *testing.T) {
-	endpoint := "/test/refresh_cache"
+	endpoint := "/test/refresh_cache/set"
+	tests := []struct {
+		name         string
+		body         string
+		expectedCode int
+		expectedBody string
+	}{
+		{
+			name:         "Missing refresh cache",
+			body:         `{"publisher": "21038", "domain": "bubu8.com"}`,
+			expectedCode: http.StatusBadRequest,
+			expectedBody: `{"message":"Refresh cache value not allowed, it should be \u003c= 500 and \u003e= 1","status":"error"}`,
+		},
+
+		{
+			name:         "Wrong value refresh cache",
+			body:         `{"publisher": "21038","refresh_cache":1800, "domain": "bubu8.com"}`,
+			expectedCode: http.StatusBadRequest,
+			expectedBody: `{"message":"Refresh cache value not allowed, it should be \u003c= 500 and \u003e= 1","status":"error"}`,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			req := httptest.NewRequest("POST", endpoint, strings.NewReader(test.body))
+			req.Header.Set("Content-Type", "application/json")
+
+			resp, err := appTest.Test(req)
+			if err != nil {
+				t.Errorf("Test %s failed: %s", test.name, err)
+				return
+			}
+			defer resp.Body.Close()
+
+			if resp.StatusCode != test.expectedCode {
+				t.Errorf("Test %s failed: expected status code %d, got %d", test.name, test.expectedCode, resp.StatusCode)
+			}
+
+			bodyBytes, err := io.ReadAll(resp.Body)
+			if err != nil {
+				t.Errorf("Test %s failed: error reading response body: %s", test.name, err)
+				return
+			}
+			bodyString := strings.TrimSpace(string(bodyBytes))
+			if bodyString != test.expectedBody {
+				t.Errorf("Test %s failed: expected body %s, got %s", test.name, test.expectedBody, bodyString)
+			}
+		})
+	}
+}
+
+func TestSetRefreshCache(t *testing.T) {
+	endpoint := "/test/refresh_cache/set"
 	tests := []struct {
 		name         string
 		body         string
@@ -23,118 +73,38 @@ func TestValidateRefreshCache(t *testing.T) {
 	}{
 
 		{
-			name:         "Invalid device",
-			body:         `{"publisher":"1234", "device": "mm", "country": "US", "refresh_cache": 1, "domain": "example.com"}`,
-			expectedCode: http.StatusBadRequest,
-			expectedBody: `{"message":"Device should be in the allowed list","status":"error"}`,
-		},
-		{
-			name:         "Invalid country",
-			body:         `{"publisher": "test", "device": "tablet", "country": "USA", "refresh_cache": 1, "domain": "example.com"}`,
-			expectedCode: http.StatusBadRequest,
-			expectedBody: `{"message":"Country code must be 2 characters long and should be in the allowed list","status":"error"}`,
+			name:         "Correct values in refresh cache",
+			body:         `{"publisher": "21038","refresh_cache":180, "domain": "test.com"}`,
+			expectedCode: http.StatusOK,
+			expectedBody: `{"status":"success","message":"Refresh cache successfully created"}`,
 		},
 	}
 
 	for _, test := range tests {
-		req := httptest.NewRequest("POST", endpoint, strings.NewReader(test.body))
-		req.Header.Set("Content-Type", "application/json")
-		resp, err := appTest.Test(req)
-		if err != nil {
-			t.Errorf("Test %s failed: %s", test.name, err)
-			continue
-		}
+		t.Run(test.name, func(t *testing.T) {
+			req := httptest.NewRequest("POST", endpoint, strings.NewReader(test.body))
+			req.Header.Set("Content-Type", "application/json")
 
-		if resp.StatusCode != test.expectedCode {
-			t.Errorf("Test %s failed: expected status code %d, got %d", test.name, test.expectedCode, resp.StatusCode)
-		}
-
-		bodyBytes, _ := io.ReadAll(resp.Body)
-		bodyString := strings.TrimSpace(string(bodyBytes))
-		if bodyString != test.expectedBody {
-			t.Errorf("Test %s failed: expected body %s, got %s", test.name, test.expectedBody, bodyString)
-		}
-	}
-}
-
-func TestCreateRefreshCacheMetadataGeneration(t *testing.T) {
-	tests := []struct {
-		name         string
-		modBC        models.RefreshCacheSlice
-		finalRules   []core.RefreshCacheRealtimeRecord
-		expectedJSON string
-	}{
-		{
-			name: "Sort By Correct Order",
-			modBC: models.RefreshCacheSlice{
-				{
-					RuleID:       "",
-					Publisher:    "20814",
-					Domain:       "stream-together.org",
-					Device:       null.StringFrom("mobile"),
-					RefreshCache: 12,
-				},
-				{
-					RuleID:       "",
-					Publisher:    "20814",
-					Domain:       "stream-together.org",
-					Device:       null.StringFrom("mobile"),
-					Country:      null.StringFrom("il"),
-					RefreshCache: 11,
-				},
-				{
-					RuleID:       "",
-					Publisher:    "20814",
-					Domain:       "stream-together.org",
-					Device:       null.StringFrom("mobile"),
-					Country:      null.StringFrom("us"),
-					RefreshCache: 14,
-				},
-			},
-			finalRules:   []core.RefreshCacheRealtimeRecord{},
-			expectedJSON: `{"rules":[{"rule":"(p=20814__d=stream-together.org__c=il__os=.*__dt=mobile__pt=.*__b=.*)","refresh_cache":11,"rule_id":"cc11f229-1d4a-5bd2-a6d0-5fae8c7a9bf4"},{"rule":"(p=20814__d=stream-together.org__c=us__os=.*__dt=mobile__pt=.*__b=.*)","refresh_cache":14,"rule_id":"a0d406cd-bf98-50ab-9ff2-1b314b27da65"},{"rule":"(p=20814__d=stream-together.org__c=.*__os=.*__dt=mobile__pt=.*__b=.*)","refresh_cache":12,"rule_id":"cb45cb97-5ca2-503d-9008-317dbbe26d10"}]}`,
-		},
-		{
-			name: "Device with null value",
-			modBC: models.RefreshCacheSlice{
-				{
-					RuleID:       "",
-					Publisher:    "20814",
-					Domain:       "stream-together.org",
-					Country:      null.StringFrom("us"),
-					RefreshCache: 11,
-				},
-			},
-			finalRules:   []core.RefreshCacheRealtimeRecord{},
-			expectedJSON: `{"rules": [{"rule": "(p=20814__d=stream-together.org__c=us__os=.*__dt=.*__pt=.*__b=.*)", "refresh_cache": 11, "rule_id": "ad18394a-ee20-58c2-bb9b-dd459550a9f7"}]}`,
-		},
-		{
-			name: "Same ruleId different input refresh cache",
-			modBC: models.RefreshCacheSlice{
-				{
-					RuleID:       "",
-					Publisher:    "20814",
-					Domain:       "stream-together.org",
-					Country:      null.StringFrom("us"),
-					Device:       null.StringFrom("mobile"),
-					RefreshCache: 14,
-				},
-			},
-			finalRules:   []core.RefreshCacheRealtimeRecord{},
-			expectedJSON: `{"rules": [{"rule": "(p=20814__d=stream-together.org__c=us__os=.*__dt=mobile__pt=.*__b=.*)", "refresh_cache": 14, "rule_id": "a0d406cd-bf98-50ab-9ff2-1b314b27da65"}]}`,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := core.CreateRefreshCacheMetadata(tt.modBC, tt.finalRules)
-
-			resultJSON, err := json.Marshal(map[string]interface{}{"rules": result})
+			resp, err := appTest.Test(req)
 			if err != nil {
-				t.Fatalf("Failed to marshal result to JSON: %v", err)
+				t.Errorf("Test %s failed: %s", test.name, err)
+				return
+			}
+			defer resp.Body.Close()
+
+			if resp.StatusCode != test.expectedCode {
+				t.Errorf("Test %s failed: expected status code %d, got %d", test.name, test.expectedCode, resp.StatusCode)
 			}
 
-			assert.JSONEq(t, tt.expectedJSON, string(resultJSON))
+			bodyBytes, err := io.ReadAll(resp.Body)
+			if err != nil {
+				t.Errorf("Test %s failed: error reading response body: %s", test.name, err)
+				return
+			}
+			bodyString := strings.TrimSpace(string(bodyBytes))
+			if bodyString != test.expectedBody {
+				t.Errorf("Test %s failed: expected body %s, got %s", test.name, test.expectedBody, bodyString)
+			}
 		})
 	}
 }
@@ -143,7 +113,7 @@ func Test_LR_ToModel(t *testing.T) {
 	t.Parallel()
 
 	type args struct {
-		refreshCache *core.RefreshCache
+		refreshCache *dto.RefreshCache
 	}
 
 	tests := []struct {
@@ -154,82 +124,52 @@ func Test_LR_ToModel(t *testing.T) {
 		{
 			name: "All fields populated",
 			args: args{
-				refreshCache: &core.RefreshCache{
-					RuleId:        "50afedac-d41a-53b0-a922-2c64c6e80623",
-					Publisher:     "Publisher1",
-					Domain:        "example.com",
-					RefreshCache:  1,
-					OS:            "Windows",
-					Country:       "US",
-					Device:        "Desktop",
-					PlacementType: "Banner",
-					Browser:       "Chrome",
+				refreshCache: &dto.RefreshCache{
+					RuleId:       "50afedac-d41a-53b0-a922-2c64c6e80623",
+					Publisher:    "Publisher1",
+					Domain:       "example.com",
+					RefreshCache: 1,
 				},
 			},
 			expected: &models.RefreshCache{
-				RuleID:        "50afedac-d41a-53b0-a922-2c64c6e80623",
-				Publisher:     "Publisher1",
-				Domain:        "example.com",
-				RefreshCache:  1,
-				Country:       null.StringFrom("US"),
-				Os:            null.StringFrom("Windows"),
-				Device:        null.StringFrom("Desktop"),
-				PlacementType: null.StringFrom("Banner"),
-				Browser:       null.StringFrom("Chrome"),
+				RuleID:       "50afedac-d41a-53b0-a922-2c64c6e80623",
+				Publisher:    "Publisher1",
+				Domain:       "example.com",
+				RefreshCache: 1,
 			},
 		},
 		{
 			name: "Some fields empty",
 			args: args{
-				refreshCache: &core.RefreshCache{
-					RuleId:        "d823a92a-83e5-5c2b-a067-b982d6cdfaf8",
-					Publisher:     "Publisher2",
-					Domain:        "example.org",
-					RefreshCache:  1,
-					OS:            "",
-					Country:       "CA",
-					Device:        "",
-					PlacementType: "Sidebar",
-					Browser:       "",
+				refreshCache: &dto.RefreshCache{
+					RuleId:       "d823a92a-83e5-5c2b-a067-b982d6cdfaf8",
+					Publisher:    "Publisher2",
+					Domain:       "example.org",
+					RefreshCache: 1,
 				},
 			},
 			expected: &models.RefreshCache{
-				RuleID:        "d823a92a-83e5-5c2b-a067-b982d6cdfaf8",
-				Publisher:     "Publisher2",
-				Domain:        "example.org",
-				RefreshCache:  1,
-				Country:       null.StringFrom("CA"),
-				Os:            null.String{},
-				Device:        null.String{},
-				PlacementType: null.StringFrom("Sidebar"),
-				Browser:       null.String{},
+				RuleID:       "d823a92a-83e5-5c2b-a067-b982d6cdfaf8",
+				Publisher:    "Publisher2",
+				Domain:       "example.org",
+				RefreshCache: 1,
 			},
 		},
 		{
 			name: "All fields empty",
 			args: args{
-				refreshCache: &core.RefreshCache{
-					RuleId:        "966affd7-d087-57a2-baff-55b926f4c32d",
-					Publisher:     "",
-					Domain:        "",
-					RefreshCache:  1,
-					OS:            "",
-					Country:       "",
-					Device:        "",
-					PlacementType: "",
-					Browser:       "",
+				refreshCache: &dto.RefreshCache{
+					RuleId:       "966affd7-d087-57a2-baff-55b926f4c32d",
+					Publisher:    "",
+					Domain:       "",
+					RefreshCache: 1,
 				},
 			},
 			expected: &models.RefreshCache{
-				RuleID:        "966affd7-d087-57a2-baff-55b926f4c32d",
-				Publisher:     "",
-				Domain:        "",
-				RefreshCache:  1,
-				Country:       null.String{},
-				Os:            null.String{},
-				Device:        null.String{},
-				PlacementType: null.String{},
-				Browser:       null.String{},
+				RuleID:       "966affd7-d087-57a2-baff-55b926f4c32d",
+				Publisher:    "",
+				Domain:       "",
+				RefreshCache: 1,
 			},
 		},
 	}
